Tryied to fix ambiguous pow() call on mac.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/async13@5699 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: f54b3ead7e45912bee4bec0fb9c4f17de8b0c505

+++ ksolve/RateTerm.h
@@ -487,7 +487,7 @@ class NOrder: public ZeroOrder
 			assert( v_.size() > 0 );
 			if ( v_.size() == 1 )
 				return new NOrder( k_, v_ );
-			return new NOrder( k_ * pow( ratio, v_.size() -1 ), v_ );
+			return new NOrder( k_ * pow( ratio, (int)(v_.size() -1) ), v_ );
 		}
 
 	protected:
@@ -513,7 +513,7 @@ class StochNOrder: public NOrder
 			assert( v_.size() > 0 );
 			if ( v_.size() == 1 )
 				return new StochNOrder( k_, v_ );
-			return new StochNOrder( k_ * pow( ratio, v_.size() -1 ), v_ );
+			return new StochNOrder( k_ * pow( ratio, (int)(v_.size() -1) ), v_ );
 		}
 };
